rename analysis.js file
