+++ packages/js/src/classic-editor/getAnalysisConfiguration.js
